New version compatible with the Matlab 2020 release
The matlab 2020 release has a built in function called 'bvar' which causes a crash in the version v4.1 of the toolbox. In the version v4.2, we create the function 'bvar_.m' and substitute 'bvar.m' with 'bvar_.m' in all the examples of the tutorial. Codes are backwarad compatible so that for matlab releases earlier than 2020 'bvar.m' still works.