Shocks and exogenous vars decomp
Historical decomposition of variables is computed also for VARX models. In this case the components of the decomposition of the endogenous variables are (1) stochastic part (shocks), (2) exogenous part (exo or control vars) and (3) the pure deterministic part. They are stored in this specific order.
example_6_VARX.m has been extended to provide an example of such decomposition.