Added VAR estimates with bootstrap methods
Added a new function cvar_.m that estimates the parameters of the VAR model with classical methods. Impulse Response Function (and forecasts) CI are computed using bootstrap methods (sampling with replacement of the VAR residuals).
Users can apply the same identification schemes as for the bvar_.m as well as a number of similar options.

An example of the use of cvar_.m is presented in the tutorial at the end of example_1_classical.m