Updates on sign/narrative restrictions algorithm
Until a correct  rotation is found, to verify the sign and the narrative restrictions the impulse responses are computed only for the (max) horizon of the sign restrictions.
once and if the rotation is found, the IRF are the computed for the desired horizon options.hor. this speed up somehow the codes in particular for the narrative restriction case.