Error in constructing the pooled VAR estimator
There was an error in constructing the pooled VAR estimator. Kudos to Hee So Kim for spotting the inconsistency. The matrix of the LHS regressors (vax.X) -used to consturct the posterior- was not reflecting the correct lag  structure of the endogenous variables.

In order to compute the pooled estimator across different units, the user needs to input the data file as a three dimension array, where the first dimension is time, the second dimension is the variable and the third dimension is the unit (i.e. country, firms ... ).

the example in the BVAR tutorial has changed accordingly.